+++ lib/core/message/scheduled_file_message.g.dart
@@ -70,7 +70,7 @@ ScheduledFileMessage _$ScheduledFileMessageFromJson(
               ?.map((e) => Reaction.fromJson(e as Map<String, dynamic>))
               .toList() ??
           []
-      ..messageTemplate =
+      ..extendedMessage =
           json['extended_message'] as Map<String, dynamic>? ?? {}
       ..requireAuth = json['require_auth'] as bool?;
 
@@ -96,7 +96,7 @@ Map<String, dynamic> _$ScheduledFileMessageToJson(
       'data': instance.data,
       'reactions': instance.reactions?.map((e) => e.toJson()).toList(),
       'scheduled_info': instance.scheduledInfo?.toJson(),
-      'extended_message': instance.messageTemplate,
+      'extended_message': instance.extendedMessage,
       'scheduled_message_id': instance.scheduledMessageId,
       'type': _$MessageTypeEnumMap[instance.type],
       'file': instance.file.toJson(),